+++ chaoscloud/api/safeguard.py
@@ -1,14 +1,20 @@
 # -*- coding: utf-8 -*-
-from typing import List, NoReturn, Optional
+from copy import deepcopy
+import threading
+
+from typing import List, NoReturn
 
 from chaoslib.exceptions import InterruptExecution
-from chaoslib.types import Extension
+from chaoslib.types import Extension, Journal
 import requests
 
-from . import urls
+from . import urls, get_chaosiq_extension_from_journal, get_execution_id, \
+    get_experiment_id
 
 
-__all__ = ["is_allowed_to_continue"]
+__all__ = ["is_allowed_to_continue", "set_applied_safeguards_for_execution"]
+safeguards_state = {}
+state_lock = threading.Lock()
 
 
 def is_allowed_to_continue(session: requests.Session,
@@ -35,24 +41,21 @@ def is_allowed_to_continue(session: requests.Session,
     state = r.json()
     if state.get("allowed", True) is False:
         safeguards = "\n".join([p["name"] for p in state.get("policies")])
+        with state_lock:
+            safeguards_state[execution_id] = deepcopy(state.get("policies"))
+
         raise InterruptExecution(
             "The following safe guards disallow this execution from "
             "continuing:\n{}".format(safeguards)
         )
 
 
-###############################################################################
-# Internals
-###############################################################################
-def get_experiment_id(extensions: List[Extension]) -> Optional[str]:
-    if not extensions:
-        return
-    for extension in extensions:
-        if extension["name"] == "chaosiq":
-            return extension["experiment_id"]
-
-
-def get_execution_id(extensions: List[Extension]) -> Optional[str]:
-    for extension in extensions:
-        if extension["name"] == "chaosiq":
-            return extension["execution_id"]
+def set_applied_safeguards_for_execution(extensions: List[Extension],
+                                         journal: Journal):
+    """
+    Set the list of safeguards applied to an execution's journal.
+    """
+    execution_id = get_execution_id(extensions)
+    extension = get_chaosiq_extension_from_journal(journal)
+    with state_lock:
+        extension["safeguards"] = safeguards_state.get(execution_id)