+++ tests/conftest.py
@@ -1,4 +1,4 @@
-from typing import Dict, List
+from typing import Any, Dict, List
 import uuid
 
 import pytest
@@ -21,3 +21,18 @@ def default_org(default_org_id: str) -> Dict[str, str]:
 @pytest.fixture
 def organizations(default_org: Dict[str, str]) -> List[Dict[str, str]]:
     return [default_org]
+
+
+@pytest.fixture
+def settings() -> Dict[str, Any]:
+    return {
+        "controls":
+            {
+                "chaosiq-cloud": {
+                    "features": {
+                        "publish": 'on',
+                        "safeguards": 'on'
+                    }
+                }
+            }
+    }